Refactor: Change created_by column type to unsignedInteger in categories table migration

+++ laravel-pss-app/database/migrations/2024_11_06_160119_create_categories_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name', 100);
             $table->text('description')->nullable();
-            $table->unsignedBigInteger('created_by');
+            $table->unsignedInteger('created_by');
             $table->foreign('created_by')->references('id')->on('admins');
             $table->timestamps();
         });
